Last

+++ c4prj3_finish/main.c
@@ -65,9 +65,9 @@ int main(int argc, char ** argv) {
 	}
       win[index]++;
       highest=NULL;
-      for(size_t k=0;k<n_hands;k++)
+      for(int k=0;k<n_hands;k++)
 	{
-	  printf("Hand %zu won %u / %u times (%.2f%%)\n",k,win[k],num_trials,(win[k]*100)/(double)num_trials);
+	  printf("Hand %d won %u / %u times (%.2f%%)\n",k,win[k],num_trials,(win[k]*100)/(double)num_trials);
 	}
       printf("And there were %u ties\n",win[n_hands]);
     }
